squash all fixes
Major changes to the core. The previous source had fundamental
problems and I gave up trying to solve them  after several tries.
It was anyway an arbitrary requirement that the forth is 32-bit
regardless of the architecture, and then additionally it was
getting complicated to make it work with an second indirection to
actual function pointer. In the end I could not make it work at
all for 2 or 3 level forth words when using the kind of unusual
forth implementation.

Revisiting the core and implementing a more standard direct
threaded code forth worked out much better, with also possible
performance gains etc.